refactoring: moved components to 'secure' folder
