+++ app/fragments/secure/components/TransferBatch.tsx
@@ -2,41 +2,41 @@ import BN from "bn.js";
 import React from "react";
 import { Alert, Platform, View, Text, ScrollView, Pressable } from "react-native";
 import { Address, Cell, CellMessage, CommonMessageInfo, ExternalMessage, fromNano, InternalMessage, SendMode, StateInit, toNano } from "ton";
-import { MixpanelEvent, trackEvent } from "../analytics/mixpanel";
-import { AppConfig } from "../AppConfig";
-import { contractFromPublicKey } from "../engine/contractFromPublicKey";
-import { useEngine } from "../engine/Engine";
-import { ContractMetadata } from "../engine/metadata/Metadata";
-import { useItem } from "../engine/persistence/PersistedItem";
-import { usePrice } from "../engine/PriceContext";
-import { JettonMasterState } from "../engine/sync/startJettonMasterSync";
-import { parseMessageBody } from "../engine/transactions/parseMessageBody";
-import { parseBody } from "../engine/transactions/parseWalletTransaction";
-import { resolveOperation } from "../engine/transactions/resolveOperation";
-import { createWalletTransferV4, internalFromSignRawMessage } from "../engine/utils/createWalletTransferV4";
-import { LocalizedResources } from "../i18n/schema";
-import { t } from "../i18n/t";
-import { KnownWallet, KnownWallets } from "../secure/KnownWallets";
-import { getCurrentAddress } from "../storage/appState";
-import { loadWalletKeys, WalletKeys } from "../storage/walletKeys";
-import { Theme } from "../Theme";
-import { warn } from "../utils/log";
-import { backoff } from "../utils/time";
-import { useTypedNavigation } from "../utils/useTypedNavigation";
-import { ItemAddress } from "./ItemAddress";
-import { ItemCollapsible } from "./ItemCollapsible";
-import { ItemDivider } from "./ItemDivider";
-import { ItemGroup } from "./ItemGroup";
-import { ItemLarge } from "./ItemLarge";
-import { PriceComponent } from "./PriceComponent";
-import { RoundButton } from "./RoundButton";
-import { TransferComponent } from "./transactions/TransferComponent";
-import { WImage } from "./WImage";
+import { MixpanelEvent, trackEvent } from "../../../analytics/mixpanel";
+import { AppConfig } from "../../../AppConfig";
+import { contractFromPublicKey } from "../../../engine/contractFromPublicKey";
+import { useEngine } from "../../../engine/Engine";
+import { ContractMetadata } from "../../../engine/metadata/Metadata";
+import { useItem } from "../../../engine/persistence/PersistedItem";
+import { usePrice } from "../../../engine/PriceContext";
+import { JettonMasterState } from "../../../engine/sync/startJettonMasterSync";
+import { parseMessageBody } from "../../../engine/transactions/parseMessageBody";
+import { parseBody } from "../../../engine/transactions/parseWalletTransaction";
+import { resolveOperation } from "../../../engine/transactions/resolveOperation";
+import { createWalletTransferV4, internalFromSignRawMessage } from "../../../engine/utils/createWalletTransferV4";
+import { LocalizedResources } from "../../../i18n/schema";
+import { t } from "../../../i18n/t";
+import { KnownWallet, KnownWallets } from "../../../secure/KnownWallets";
+import { getCurrentAddress } from "../../../storage/appState";
+import { loadWalletKeys, WalletKeys } from "../../../storage/walletKeys";
+import { Theme } from "../../../Theme";
+import { warn } from "../../../utils/log";
+import { backoff } from "../../../utils/time";
+import { useTypedNavigation } from "../../../utils/useTypedNavigation";
+import { ItemAddress } from "../../../components/ItemAddress";
+import { ItemCollapsible } from "../../../components/ItemCollapsible";
+import { ItemDivider } from "../../../components/ItemDivider";
+import { ItemGroup } from "../../../components/ItemGroup";
+import { ItemLarge } from "../../../components/ItemLarge";
+import { PriceComponent } from "../../../components/PriceComponent";
+import { RoundButton } from "../../../components/RoundButton";
+import { TransferComponent } from "../../../components/transactions/TransferComponent";
+import { WImage } from "../../../components/WImage";
 import Question from '../../assets/ic_question.svg';
 import TonSign from '../../assets/ic_ton_sign.svg';
 import LottieView from 'lottie-react-native';
 import SignLock from '../../assets/ic_sign_lock.svg';
-import { formatCurrency } from "../utils/formatCurrency";
+import { formatCurrency } from "../../../utils/formatCurrency";
 
 type Props = {
     text: string | null,
